requirements.txt
praw==7.7.1
pymongo==4.6.0
pandas==2.1.3
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
matplotlib==3.8.2
plotly==5.18.0
streamlit==1.29.0
wordcloud==1.9.3
python-dotenv==1.0.0
dnspython==2.4.2
requests==2.31.0
numpy==1.26.2
Pillow==10.1.0
